Change theme component exports to named exports
This is done so that component names in React inspector match
the override names. To accomplish that, a named export with variable
matching the filename is exported instead of default.

Otherwise, the only way to do this would be to use variable shadowing, like:
```ts
export const EmailPasswordSignUp = withOverride("EmailPasswordSignUp", class EmailPasswordSignUp extends ... );
// or
export const EmailPasswordSignUp = withOverride("EmailPasswordSignUp", function EmailPasswordSignUp(...));
```

Since variable shadowing is not recommended and this is pretty ugly, unexpected and "tricky", current approach was used.